Add tasks.h file
